requirements.txt
# Core deep learning
torch>=2.2.0
torchvision>=0.17.0
monai>=1.3.0

# Data processing
pandas
numpy
scikit-learn
pillow
pydicom

# Indexing
faiss-cpu

# Interface
gradio>=4.24.0
